WIP jpa aggregation.

+++ src/main/java/org/tdl/vireo/model/repo/impl/WorkflowStepRepoImpl.java
@@ -15,8 +15,8 @@ import org.tdl.vireo.model.repo.WorkflowStepRepo;
 import org.tdl.vireo.model.repo.custom.WorkflowStepRepoCustom;
 
 public class WorkflowStepRepoImpl implements WorkflowStepRepoCustom {
-    
-    @PersistenceContext
+	
+	@PersistenceContext
     private EntityManager em;
 	
     @Autowired
@@ -32,33 +32,45 @@ public class WorkflowStepRepoImpl implements WorkflowStepRepoCustom {
     public WorkflowStep create(String name, Organization originatingOrganization) {
         WorkflowStep workflowStep = workflowStepRepo.save(new WorkflowStep(name, originatingOrganization));
         originatingOrganization.addWorkflowStep(workflowStep);
+        organizationRepo.save(originatingOrganization);
         return workflowStep;
     }
     
     // TODO: this method needs to handle all inheretence and aggregation duties
     public WorkflowStep update(WorkflowStep workflowStep, Organization requestingOrganization) {
-        
+    	
         Organization originatingOrganization = workflowStep.getOriginatingOrganization();
-        WorkflowStep originatingWorkflowStep = workflowStep.getOriginatingWorkflowStep();
         
         // If the requestingOrganization originates the workflowStep, make the change directly
         if(originatingOrganization != null && requestingOrganization.getId().equals(originatingOrganization.getId())) {
             
             if(!workflowStep.getOverrideable()) {
-                            	
-            	List<WorkflowStep> derivativeWorkflowStepsToDelete = getDescendantsOfStep(workflowStep);
                 
-                for(WorkflowStep ws : derivativeWorkflowStepsToDelete){
+                List<WorkflowStep> descendents = getDescendantsOfStep(workflowStep);
+                
+                for(WorkflowStep ws : descendents) {
                     
-                    organizationRepo.findByWorkflowId(ws.getId()).forEach(organization -> {
-                        //TODO:  this puts it at the end; we should make sure it gets into the right spot
-                        organization.addStepToWorkflow(originatingWorkflowStep);
-                        organizationRepo.save(organization);
-                    });
+                    System.out.println("descendent " + ws.getId());
                     
+                    for(Organization organization : organizationRepo.findByWorkflowId(ws.getId())) {
+                        
+                        System.out.println("  org " + organization.getId() + " " + organization.getName());
+                        System.out.println("    replacing " + ws.getId() + " with " + workflowStep.getId());
+                        System.out.println("    replaced: " + organization.replaceStepInWorkflow(ws, workflowStep));
+                        System.out.println("    replaced: " + organization.replaceWorkflowStep(ws, workflowStep));
+                        
+                        organization = organizationRepo.save(organization);
+                        
+                        organization.getWorkflow().forEach(tws -> {
+                            System.out.println("                      in replace " + tws.getId());
+                        });
+                    }
+                    
+                }
+                
+                for(WorkflowStep ws : descendents) {
                     workflowStepRepo.delete(ws);
-                    
-                } 
+                }
                 
             }
             
@@ -68,10 +80,10 @@ public class WorkflowStepRepoImpl implements WorkflowStepRepoCustom {
         	
             if(workflowStep.getOverrideable()) {
             	
-            	Long originalWorkflowStepId = workflowStep.getId();        
-                
-                em.detach(workflowStep);
-                
+            	
+            	Long originalWorkflowStepId = workflowStep.getId();
+            	
+            	em.detach(workflowStep);
                 workflowStep.setId(null);
                 
                 workflowStep = workflowStepRepo.save(workflowStep);
@@ -79,11 +91,24 @@ public class WorkflowStepRepoImpl implements WorkflowStepRepoCustom {
                 
                 WorkflowStep originalWorkflowStep = workflowStepRepo.findOne(originalWorkflowStepId);
                 
-            	workflowStep.setOriginatingWorkflowStep(originalWorkflowStep);
-            	workflowStep.setOriginatingOrganization(requestingOrganization);                
+                requestingOrganization.removeWorkflowStep(originalWorkflowStep);
+                                
+                System.out.println("new: " + workflowStep.getId());
+                System.out.println("original: " + originalWorkflowStep.getId());
                 
-            	requestingOrganization.addWorkflowStep(workflowStep);
+                workflowStep.setOriginatingWorkflowStep(originalWorkflowStep);
+                workflowStep.setOriginatingOrganization(requestingOrganization);
+                
+                
+                if(requestingOrganization.getWorkflow().contains(originalWorkflowStep)) {
+                    requestingOrganization.replaceStepInWorkflow(originalWorkflowStep, workflowStep);
+                }
+                else {
+                    requestingOrganization.addWorkflowStep(workflowStep);
+                }
+                
             	organizationRepo.save(requestingOrganization);
+            	            	
             }
             else {
             	
@@ -95,61 +120,59 @@ public class WorkflowStepRepoImpl implements WorkflowStepRepoCustom {
             
         }
         
-        
-        
-        
         return workflowStepRepo.save(workflowStep);
     }
     
     @Override
     public void delete(WorkflowStep workflowStep) {
-    	
-        Organization originatingOrganization = workflowStep.getOriginatingOrganization();
         
-        originatingOrganization.removeWorkflowStep(workflowStep);
+        if(workflowStep != null && workflowStepRepo.findOne(workflowStep.getId()) != null) {
+
+            Organization originatingOrganization = workflowStep.getOriginatingOrganization();
+            
+            originatingOrganization.removeWorkflowStep(workflowStep);
+                    
+            if(workflowStep.getOriginatingWorkflowStep() != null) {
+                workflowStep.setOriginatingWorkflowStep(null);
+            }
+            
+            fieldProfileRepo.findByOriginatingWorkflowStep(workflowStep).forEach(fieldProfile -> {
+                workflowStep.removeFieldProfile(fieldProfile);
+                workflowStepRepo.save(workflowStep);
+                fieldProfileRepo.delete(fieldProfile);
+            });
+            
+            organizationRepo.findByWorkflowId(workflowStep.getId()).forEach(organization -> {
+                organization.removeStepFromWorkflow(workflowStep);
+                organizationRepo.save(organization);
+            });
+            
+            workflowStepRepo.findByOriginatingWorkflowStep(workflowStep).forEach(ws -> {
+                ws.setOriginatingWorkflowStep(null);
+            });
+            
+            deleteDescendantsOfStep(workflowStep);
+            
+            workflowStepRepo.delete(workflowStep.getId());
         
-        organizationRepo.save(originatingOrganization);
-        
-        if(workflowStep.getOriginatingWorkflowStep() != null) {
-        	workflowStep.setOriginatingWorkflowStep(null);
         }
-        
-        fieldProfileRepo.findByOriginatingWorkflowStep(workflowStep).forEach(fieldProfile -> {
-        	workflowStep.removeFieldProfile(fieldProfile);
-        	workflowStepRepo.save(workflowStep);
-        	fieldProfileRepo.delete(fieldProfile);
-        });
-         
-        organizationRepo.findByWorkflowId(workflowStep.getId()).forEach(organization -> {
-        	organization.removeWorkflowStep(workflowStep);
-            organizationRepo.save(organization);
-        });
-        
-        workflowStepRepo.findByOriginatingWorkflowStep(workflowStep).forEach(ws -> {
-        	ws.setOriginatingWorkflowStep(null);
-        });
-        
-        deleteDescendantsOfStep(workflowStep);
-        
-        workflowStepRepo.delete(workflowStep.getId());
     }
     
     private void deleteDescendantsOfStep(WorkflowStep workflowStep) {
         workflowStepRepo.findByOriginatingWorkflowStep(workflowStep).forEach(desendantWorflowStep -> {
-        	deleteDescendantsOfStep(desendantWorflowStep);
-        	delete(desendantWorflowStep);
+            deleteDescendantsOfStep(desendantWorflowStep);
+            delete(desendantWorflowStep);
         });
     }
     
     private List<WorkflowStep> getDescendantsOfStep(WorkflowStep workflowStep) {
         List<WorkflowStep> descendantWorkflowSteps = new ArrayList<WorkflowStep>();
-        List<WorkflowStep> currentWorkflowSteps = workflowStepRepo.findByOriginatingWorkflowStep(workflowStep);
-        descendantWorkflowSteps.addAll(currentWorkflowSteps);
-        currentWorkflowSteps.forEach(desendantWorflowStep -> {
+        List<WorkflowStep> currentDescendentsWorkflowSteps = workflowStepRepo.findByOriginatingWorkflowStep(workflowStep);
+        descendantWorkflowSteps.addAll(currentDescendentsWorkflowSteps);
+        currentDescendentsWorkflowSteps.forEach(desendantWorflowStep -> {
             descendantWorkflowSteps.addAll(getDescendantsOfStep(desendantWorflowStep));
         });
         return descendantWorkflowSteps;
     }
-
     
 }
